+++ client/src/components/SwampStudy.tsx
@@ -1,0 +1,10 @@
+// The SwampStudy logo text
+export default function SwampStudy() {
+  return (
+    <span>
+      <span style={{color: "#0021A5", fontWeight: "bold"}}>Swamp</span>
+      <span style={{color: "#FA4616", fontWeight: "bold"}}>Study</span>
+    </span>
+  );
+}
+
